requirements.txt
# Web框架
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# 音频处理
librosa>=0.10.1
pydub>=0.25.1
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0

# AI和机器学习
langchain>=0.0.350
openai>=1.3.0
transformers>=4.35.0

# 数据处理
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.13.0

# HTTP和API
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# 文件处理
python-multipart>=0.0.6
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.5.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 开发工具
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
